backend/requirements.txt
Django==3.2.3
django-filter==23.2
djangorestframework==3.12.4
djoser==2.1.0
python-dotenv==1.0.0
django-extensions==3.2.3
PyYAML==6.0
Pillow==10.1.0
gunicorn==20.1.0
reportlab==4.0.7